Initial commit.
The project layout was generated using cookiecutter with the template
at https://github.com/wildfish/wildfish-django-starter/ from the
'upgrade' branch.

There are a few things that are probably superfluous for the National
Screening Centre project but these were left in until the initial
deployment of the site for user testing becomes clear - the target
platform is the PHE cloud but we might deploy initially to a droplet
just to get up and running quickly.

The final configuration will crystalise once the project gets up and
running and the excess features here will be removed.