Renamed the initial policy app to core.
When the project was generated an initial app named policy was created
along with a Policy model and associated views. That was enough to check
out the basic project generated by cookiecutter.

Now the project was restructured to allow different models to be added
so the app was renamed from 'policy' to 'core' and the model renamed
to 'Condition' with the views, templates, etc. updated accordingly.

The project URLs were also updated to add namespaces for each model,
i.e. condition. That makes it easier to add new views for the other
models that will be need for this project and keep the code
sufficiently modular to make changes easy.