Refactored scripts to initial models from legacy site.
The previous script, generate_fixtures created the fixtures files for both
the Condition and Policy models since it was derived from the same table
on PHE's current (lagacy) recommendation site and there's a foreign key
from Policy to Condition. The decision was taken to simply the script
and split it into two, one for conditions and one for policies. In future
as more scripts are added or the exising ones expanded it should be easier
to understand what's going on.