Refactored the core app.
Instead of having a central app which would contain the models and views
for the application instead everything will be split into separate apps
to keep things nice and simple.

The core app contains the Condition and Policy models and related views.
These will be reincarnated in separate apps. it was easier from a code
review perspective to simply delete everything in core and re-add as
necessary.