Re-added the part of the core app that contained the Condition model.
The condition app is stripped down so it only contains the model,
django admin and related code. The view and related code were removed
as it is not clear that viewing and editing instances of the Condition
model are going to be needed right now. However there is condition
specific content - the description of the condition and the link to
more information on the main National Health Service web site, so it
seems likely that views will be added in the future.