HTML descriptions on Condition and Policy models are generated from markdown.
Added markdown (https://github.com/Python-Markdown/markdown) to generate the
HTML that will be displayed on a page when a Condition and Policy model is
saved. The attr_list extension allows the write to add the css classes that
ensure the rendered text follows the GDS style.

It's not clear if explicitly annotating the markdown using attribute lists
will work for the content managers. However since they supposedly already
write content in markdown we will run with this option until a better one
is needed.