Condition app is now used to display public pages for policies.
The policy pages have two audiences: a set of pages where the public can
see the list of conditions that have a policy of screen or not screen and
a set of admin pages where the content of the policy pages can be edited.

That created a problem since there were two lists of policies and two
details pages details pages describing a policy - one viewed by the public
and one used by admins. That raises the issue of how to organise the code
to keep the structure simple and easy to follow.

The merging of the Condition fields into the Policy model essentially
left the condition app without a specific purpose so it was decided that
instead of splitting the policy app code into two: one for admin pages
and one for public pages, the public pages would be moved over to the
condition app. That kept the code in each clean, simple and easy to
understand.