Removed admin and model utils and updated admin tests.
Now that the scope of the project is clearer the final number of
models and hence Django admin pages is likely to be low. As a result
the tests for each set of views in the Django Admin are added to each
app explicitly rather than having a generic set of tests for all admin
pages based on the model and admin utility functions added initially.

The utility functions added for querying project information on apps,
models and the Django admin are unlikey to be used extensively, if at
all, so they were removed.