Added a block to the base template for django messages.
Add a block for displaying the list of messages from django.contrib.messages.
The HTML just has basic scss styles (there was too much to do it in a style
attribute). A proper set of styles will need to be created, particularly to
support the different message levels. This is left for a later task since it
needs input from the Public Health England Design Team - a ticket was created
for this.