Moved site-wide static assets to frontend/
The path static/src/ doesn't need to be made public, and the files in
static/dist/ would make more sense being served from /static/. This is
addressed by changing STATICFILES_DIRS in settings.py.

At that point the static/ dir name is confusing, as it implies it will
be included in collectstatic. Renaming it to frontend makes its purpose
clearer.