Add app for managing stakeholders.
Most stakeholders are interested in a small number of conditions and
typically have one or two people who are the contacts for
notifications of when the conditions they are interested in come up for
review.

Two apps were created, organisation and contact with corresponding
models. Contacts are associated with the organisation. All the Policy
(conditions) the stakeholder is interested in are also associated with
the Organisation. The goal was to create a design that was
sufficiently simple and met the needs of the majority of
stakeholders. For smaller organisations the contacts are likely
directly interested in the set of conditions. In larger organisations
the contacts will likely be administration staff notifying their staff
whenever a notification of a review is received from PHE.