The organisation to add a new contact to is now specified in the url.
After an organisation was created the contact was only ever added when
reviewing the organisation details. Rather than pass slugs or pks in
the URL the value was set in the user's session. That, of course,
leads to problems if the user ever opens browser tabs to two different
organisations. Now the organisation primary key is passed in the URL,
that's a lot shorter thanthe slug (which can be up to 70 characters).
The general policy is to use slugs as that makes for more readable
URLs (so user's can see where they are) however in this case it's
unlikely a user would bookmarked the "add contact" page so usnig a
primary key is not going to generate undue confusion.