Added pages for members of the public to submit comments on a review.
The commit in intended simply to get the code up to a level similar to
the beta version so we can start to have a dialog with the designers.
The pages are visually complete, more or less, but lack functionality,
validation, tests and optimisation, particularly for the HTML. The goal
is simply to use them as a reference point so we can discuss the details
of the user stories and so be able to write tickets.