Added the workflow for managing reviews.
This is really version 0 of the screens from the prototype application
so we have something concrete to discuss with the Design team. The
models are likely close to their final implementation (might need a
little attribute shiffing) however the rest of the code should be
considered a work in progress until the results of the user experience
reviews are understood and published.