Added keyword search to the list of policies and conditions.
Added a text field to the Policy model which enables an admin to enter
keywords as free text. When a member of the public searches the list
of conditions or when an admin searches the list of policies then the
keywords field is included in the search.

The search is performed as a simple case-insensitive exact match.
PostgreSQL's full text search using SearchVector and SearchQuery were
investigated but 'icontains' gave the better results.

One unresolved issue is whether the text entered into the name field
is treated as a name or as a set of keywords. If it is the latter then
a decision will be needed on whether the keywords are AND'ed or OR'ed
together. Since the filtering is for members of the public and PHE
staff it is likely that a simple text match will be sufficient as
a single word or part of a word is the most like way this will be used.