Restored link title in case multiple links are present.
When the description of a condition was scraped from the leagacy site the
assumption was made that there would only be one link to additional
information typical on the main UK National Health Service site. As a result
the text "Read more on NHS UK" was used. it turns out that some conditions
actually have two links so the text scraped from the site was used instead.

Also when scraping the condition description extra blank lines were being
included in the description. These are now ignored.