Removed extra admin home.
The initial version of the prototype had two home pages for admins:
one at / which showed the latest status of the reviews and the other
at /admin/ which was a landing page with jump points for the various
actions that could be performed. The latter was just an early version
before any user testing was carried out and it no longer used, so it
was removed.