Added separate set of pages for public and stakeholder comments.
When submitting a comment during the public consultation period of a
review a user may submit them as a member of the public or as a
stakeholder - either an activist, charity or institution. The content
of the pages displayed varies slightly but the flow is essentially
the same - though the emails use different templates to account for
the different fields filled out.

Initially a shared form and pages were considered but it seemed that the
added complexity in the code did not seem worth it so separate urls,
forms, views and templates were used for the different flows. This
generates some duplication but the code is easier to follow. Also it
is not clear that the current contents of the template are finalised,
particularly the final page when the comment has bee submitted. It
appears that for a member of the public, unfamiliar with the web site
the next step they should take is not clear. As a result it is expected
that the content and possibly the pages displayed in the two flows
will diverge further in the future.