Moved many to many relationship to Policy.
The many t0 many relationship that linked the Policy and Review models was
moved from Review to Policy. That made it easier to construct custom
queries for fetching Policy objects and retrieving the associated reviews
depending on their various states: open for comment, under review but
closed for comment, etc.