Defaults are set when a Review is created.
Previously the slug and summary_html were set when the clean method was
called on the model. The review_start date was added to this but this
attribute needs to be set when a Review is created so it made more sense
to move all the setting of defaults or updating existing attributes to
the model's save() method.