Reworked template to make changing column widths easier.
Most of pages use a column that is two-thirds of the available width,
on the left side of the page. However ocasionally the full width is
needed for long titles or content containing tables. The "narrow"
template was modified so this can be done on an as-needed basis rather
than trying to create different templates for each case. That still
might be better but only when we've done all the pages - for now this
is simpler, easy to deal with and does not take much effort to use.