Updated pages in condition edit/preview/publish workflow.
1. Added two new text fields for the plain english summary and a section
   describing the history of reviews of the condition.

2. Updated the layout and look of the pages to match the V8 prototype.

3. Removed all messages for now until we get confirmation from the
   designers that they are useful - they were added in as a nice to
   have but it's going to create more work to finish them all so it's
   probably better to remove them for now, particularly since the
   styling is not the best.

There is one issue when the preview page is displayed (it's the same
view as the edit page except all the form fields are hidden). The back
link to return to editing the fields needs to behave as a button so the
form is submitted. There's a fragment of javascript on the link to do
the job. It should comply with accessibility requirements but that's
not known for certain.