Button shown on policy page reflects whether a review is open or not.
On the policy page seen by Evidence Review Managers if a review is
on going then the button in the top right corner shows "Manage product"
and clicking on it takes them to the review detail page where they can
complete the various tasks. If there is no currently open review then
a "Create a new product" button is displayed and clicking on it takes
them to the create review page.